requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
python-dateutil==2.8.1
requests==2.24.0
six==1.15.0
urllib3==1.25.10
Werkzeug==1.0.1
wincertstore==0.2
